% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R, R/main.R
\docType{package}
\name{gs}
\alias{gs}
\alias{gs-package}
\title{gs: A package for computating Graph Splitting Method.}
\usage{
gs(infile, type = "fst", opn = 11, ext = 1, mtx = "BLOSUM62", dup = 0,
  ...)
}
\arguments{
\item{infile}{A file name in "Multiple fasta", "BLAST output" or "Sequence similarity graph" format.
See \url{https://en.wikipedia.org/wiki/FASTA_format} for more details.}

\item{type}{A character. Each of "fst", "ml" and "mtx" specifies the file format 
as "Multiple fasta", "BLAST output" or "Sequence similarity graph", respectively.}

\item{opn}{A integer scalar. BLAST option (cost to open a gap).}

\item{ext}{A integer scalar. BLAST option (cost to extend a gap).}

\item{mtx}{A character. BLAST option (Scoring matrix name).}

\item{dup}{A integer scalar. Duplication number for the Edge Perturbation method.}
}
\description{
The gs package provides two important functions:
gs and ep.

\code{gs} reconstructs the GS tree with EP values from protein sequences.
}
\examples{
gs("tests/test.faa")
gs("tests/test.faa", dup=10)

}
\seealso{
\url{http://gs.bs.s.u-tokyo.ac.jp}
}

